+++ config/symbols.stmad.txt
@@ -20,6 +20,9 @@ g_testColliFrames = 0x8018095C;
 g_testCollPrizeTable = 0x80180968;
 g_testCollRandTable = 0x801809A8;
 g_eDamageDisplayClut = 0x801809C4;
+unk14_yVel = 0x80180FE4;
+unk14_startFrame = 0x80180FFC;
+unk14_lifetime = 0x80181000;
 g_Rooms = 0x80181130;
 g_eRedDoorUV = 0x801809D4;
 g_eRedDoorTiles = 0x80180594;
@@ -77,6 +80,7 @@ g_InitializeEntityData0 = 0x80180508;
 g_InitializeData0 = 0x801804FC;
 g_UnkPrimHelperRot = 0x8018D8C0;
 UnkPrimHelper = 0x80198FA0;
+EntityUnkId14 = 0x80195A54;
 EntityUnkId15 = 0x80195B44;
 UnkPolyFunc2 = 0x801996EC;
 PrimDecreaseBrightness = 0x8019976C;
