+++ config/symbols.us.stcen.txt
@@ -21,6 +21,9 @@ g_eDamageDisplayClut = 0x80180B94;
 g_eRedDoorUV = 0x80180BA4;
 c_GoldPrizes = 0x80180F58;
 c_HeartPrizes = 0x80180FE8;
+unk14_yVel = 0x80181148;
+unk14_startFrame = 0x80181160;
+unk14_lifetime = 0x80181164;
 g_Rooms = 0x801812D4;
 g_UnkPrimHelperRot = 0x8018D5F8;
 EntityBackgroundBlock = 0x8018D600;
@@ -61,7 +64,7 @@ EntityHeartDrop = 0x80197A00;
 EntityMessageBox = 0x80197B28;
 CheckColliderOffsets = 0x80198084;
 EntityUnkId13 = 0x80198174;
-EntityExplosion14 = 0x8019849C;
+EntityUnkId14 = 0x8019849C;
 EntityUnkId15 = 0x8019858C;
 EntityIntenseExplosion = 0x80198F2C;
 EntityStageNamePopup = 0x8019A5F0;
