+++ config/symbols.us.stdre.txt
@@ -28,6 +28,9 @@ g_eDamageDisplayClut = 0x80180D58;
 g_eRedDoorUV = 0x80180D68;
 c_GoldPrizes = 0x8018111C;
 c_HeartPrizes = 0x801811AC;
+unk14_yVel = 0x8018130C;
+unk14_startFrame = 0x80181324;
+unk14_lifetime = 0x80181328;
 g_Rooms = 0x80181498;
 g_UnkPrimHelperRot = 0x80191A5C;
 EntityUnkId11 = 0x80191A64;
@@ -73,7 +76,7 @@ EntityHeartDrop = 0x8019DB44;
 EntityMessageBox = 0x8019DC6C;
 CheckColliderOffsets = 0x8019E1C8;
 EntityUnkId13 = 0x8019E2B8;
-EntityExplosion14 = 0x8019E5E0;
+EntityUnkId14 = 0x8019E5E0;
 EntityUnkId15 = 0x8019E6D0;
 EntityIntenseExplosion = 0x8019F070;
 EntityStageNamePopup = 0x801A0734;
