+++ config/symbols.us.stno3.txt
@@ -28,6 +28,9 @@ g_eDamageDisplayClut = 0x8018209C;
 g_eRedDoorUV = 0x801820AC;
 c_GoldPrizes = 0x80182460;
 c_HeartPrizes = 0x801824F0;
+unk14_yVel = 0x80182650;
+unk14_startFrame = 0x80182668;
+unk14_lifetime = 0x8018266C;
 g_Rooms = 0x80183CC4;
 g_UnkPrimHelperRot = 0x801B77CC;
 EntityCavernDoorVase = 0x801B77D4;
