+++ config/symbols.us.stnp3.txt
@@ -28,6 +28,9 @@ g_eDamageDisplayClut = 0x80181A28;
 g_eRedDoorUV = 0x80181A38;
 c_GoldPrizes = 0x80181DEC;
 c_HeartPrizes = 0x80181E7C;
+unk14_yVel = 0x80181FDC;
+unk14_startFrame = 0x80181FF4;
+unk14_lifetime = 0x80181FF8;
 g_Rooms = 0x80183A7C;
 g_UnkPrimHelperRot = 0x801B2464;
 EntityBreakable = 0x801B26FC;
@@ -83,6 +86,7 @@ EntityRelicOrb = 0x801BF1A0;
 EntityHeartDrop = 0x801BFC70;
 EntityMessageBox = 0x801BFD98;
 EntityUnkId15Spawner = 0x801C0624;
+EntityUnkId14 = 0x801C070C;
 EntityUnkId15 = 0x801C07FC;
 EntityIntenseExplosion = 0x801C119C;
 EntityStageNamePopup = 0x801C2860;
