+++ config/symbols.us.stnz0.txt
@@ -28,6 +28,9 @@ g_eDamageDisplayClut = 0x80181950;
 g_eRedDoorUV = 0x80181960;
 c_GoldPrizes = 0x80181D14;
 c_HeartPrizes = 0x80181DA4;
+unk14_yVel = 0x80181F04;
+unk14_startFrame = 0x80181F1C;
+unk14_lifetime = 0x80181F20;
 g_Rooms = 0x8018272C;
 g_UnkPrimHelperRot = 0x801B0934;
 EntityBreakable = 0x801B0EEC;
@@ -79,6 +82,7 @@ EntityEquipItemDrop = 0x801BED20;
 EntityRelicOrb = 0x801BF5B8;
 EntityHeartDrop = 0x801C0088;
 EntityMessageBox = 0x801C01B0;
+EntityUnkId14 = 0x801C0B24;
 EntityUnkId15 = 0x801C0C14;
 EntityIntenseExplosion = 0x801C15B4;
 EntitySoulStealOrb = 0x801C2AA8;
