+++ config/symbols.us.strwrp.txt
@@ -16,6 +16,9 @@ g_testCollPrizeTable = 0x80180A10;
 g_testCollRandTable = 0x80180A50;
 c_GoldPrizes = 0x80180E30;
 c_HeartPrizes = 0x80180EC0;
+unk14_yVel = 0x80181020;
+unk14_startFrame = 0x80181038;
+unk14_lifetime = 0x8018103C;
 g_Rooms = 0x801811AC;
 g_UnkPrimHelperRot = 0x80188DE8;
 EntityBreakable = 0x8018908C;
@@ -53,6 +56,7 @@ EntityRelicOrb = 0x8019024C;
 EntityHeartDrop = 0x80190D1C;
 EntityMessageBox = 0x80190E44;
 CheckColliderOffsets = 0x801913A0;
+EntityUnkId14 = 0x801917B8;
 EntityUnkId15 = 0x801918A8;
 EntityIntenseExplosion = 0x80192248;
 EntityStageNamePopup = 0x8019390C;
