+++ config/symbols.us.stst0.txt
@@ -17,6 +17,9 @@ g_testCollElementLookup = 0x80181948;
 g_testColliFrames = 0x8018195C;
 g_eDamageDisplayClut = 0x80181968;
 g_eRedDoorUV = 0x80181978;
+unk14_yVel = 0x80181ED8;
+unk14_startFrame = 0x80181EF0;
+unk14_lifetime = 0x80181EF4;
 g_Rooms = 0x80182060;
 D_80183D1C = 0x80183D1C;
 D_80197F80 = 0x80197F80;
@@ -62,6 +65,7 @@ EntityPrizeDrop = 0x801B6358;
 EntityExplosion = 0x801B6B60;
 EntityEquipItemDrop = 0x801B6D34;
 CheckColliderOffsets = 0x801B7B0C;
+EntityUnkId14 = 0x801B7F24;
 EntityUnkId15 = 0x801B8014;
 EntityIntenseExplosion = 0x801B89B4;
 EntitySoulStealOrb = 0x801B9EA8;
