+++ src/st/mad/mad.h
@@ -59,9 +59,6 @@ extern s16 D_80180EDA[];
 extern u16 D_80180F5C[];
 extern u16 D_80180FBC[];
 extern s32 D_80180FCC[];
-extern s32 D_80180FE4[];
-extern u8 D_80180FFC[];
-extern u16 D_80181000[];
 extern ObjInit2 D_801810F4[];
 extern u16 g_InitializeEntityData0[];
 extern s16 D_801809EC[];
