+++ src/st/no3/48A84.c
@@ -54,28 +54,7 @@ void func_801C8C84(
 
 #include "../entity_unkId15_spawner.h"
 
-void EntityUnkId14(Entity* entity) {
-    if (entity->step == 0) {
-        entity->velocityY = D_80182650[entity->ext.generic.unk94];
-        entity->flags = FLAG_UNK_2000 | FLAG_UNK_04000000 | FLAG_UNK_08000000;
-        entity->palette = 0x8195;
-        entity->animSet = ANIMSET_DRA(2);
-        entity->animCurFrame = D_80182668[entity->params];
-        entity->drawMode = DRAW_TPAGE;
-        entity->step++;
-    } else {
-        entity->animFrameDuration++;
-        entity->posY.val -= entity->velocityY;
-
-        if (!(entity->animFrameDuration & 1)) {
-            entity->animCurFrame++;
-        }
-
-        if (D_8018266C[entity->params] < (s32)entity->animFrameDuration) {
-            DestroyEntity(entity);
-        }
-    }
-}
+#include "../entity_unkId14.h"
 
 void EntityUnkId15(Entity* arg0) {
     if (arg0->step == 0) {
