+++ src/st/no3/no3.h
@@ -192,9 +192,6 @@ extern s8 c_HeartPrizes[];
 extern s32 D_801824F4[];
 extern u8* D_801825A0[];
 extern u16 D_80182628[];
-extern s32 D_80182650[];
-extern u8 D_80182668[];
-extern u16 D_8018266C[];
 extern s32 D_80182638[];
 extern u8 D_8018267C[];
 extern ObjInit2 D_80182764[];
