+++ src/st/np3/np3.h
@@ -52,7 +52,7 @@ extern void func_801BB3F4(s16);
 extern s32 func_801BD308(u16* hitSensors, s16 sensorCount);
 extern s32 GetPlayerCollisionWith(Entity* entity, u16 w, u16 h, u16 flags);
 extern s32 GetDistanceToPlayerX(void);
-extern void func_801C070C(Entity* entity);
+extern void EntityUnkId14(Entity* entity);
 extern void func_801CD658();
 extern void EntityPrizeDrop(Entity* entity);
 extern void EntityEquipItemDrop(Entity* entity);
@@ -229,9 +229,6 @@ extern const u8* D_80181F2C[];
 extern u16 D_80181FA4[];
 extern s16 D_80181FB4[];
 extern u32 D_80181FC4[];
-extern s32 D_80181FDC[];
-extern u8 D_80181FF4[];
-extern u16 D_80181FF8[];
 extern s16* D_80182000;
 extern u8 D_80182008[];
 extern s32 D_80182204[];
