+++ src/st/nz0/nz0.h
@@ -43,7 +43,7 @@ s32 func_801BD720(u16* hitSensors, s16 sensorCount);
 s32 GetPlayerCollisionWith(Entity* entity, u16 w, u16 h, u16 flags);
 void EntityExplosion(Entity*);
 void func_801C33D8(const u32*, s32);
-void func_801C0B24(Entity* entity);
+void EntityUnkId14(Entity* entity);
 void EntityUnkId15(Entity* entity);
 void func_801C4CC0(void);
 
@@ -192,9 +192,6 @@ extern u8 D_80181DD0[];
 extern u8 D_80181DD0[];
 extern const u8* D_80181E54[];
 extern u16 D_80181ECC[];
-extern u8 D_80181F1C[];
-extern s32 D_80181F04[];
-extern u16 D_80181F20[];
 extern s16* D_80181F28;
 extern s32* D_80180EB8;
 extern s32 D_80180ED0[];
