+++ src/st/rwrp/113A0.c
@@ -8,7 +8,7 @@ INCLUDE_ASM("st/rwrp/nonmatchings/113A0", func_801915A0);
 
 #include "../entity_unkId15_spawner.h"
 
-INCLUDE_ASM("st/rwrp/nonmatchings/113A0", func_801917B8);
+#include "../entity_unkId14.h"
 
 INCLUDE_ASM("st/rwrp/nonmatchings/113A0", EntityUnkId15);
